+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Bower module
  - Component module
  - Composer module
+ - Git ignore module support
 
 ## [2.0.0] - 2019-03-14
 ### Added